feat: publish responsive workspace toolbar
