+++ .changeset/strong-toolbar-slots.md
@@ -1,0 +1,5 @@
+---
+"@stll/workspace-ui": minor
+---
+
+Expose the responsive action toolbar for compact workspace controls.
