+++ packages/workspace-ui/package.json
@@ -39,6 +39,7 @@
     "./kanban-grouping-picker": "./src/kanban-grouping-picker.tsx",
     "./kanban-view": "./src/kanban-view.ts",
     "./property-icon": "./src/property-icon.tsx",
+    "./responsive-action-toolbar": "./src/responsive-action-toolbar.tsx",
     "./sorts": "./src/sorts.tsx",
     "./table-skeleton-rows": "./src/table-skeleton-rows.tsx",
     "./types": "./src/types.ts",
